organizing codes initially developed for correlation analysis
